Conflicts:
	Warlord.sln.docstates.suo
	Warlord.suo
	Warlord/Warlord/Event/EventDataMap.txt
	Warlord/Warlord/Event/ValidEventInitializer.cs
	Warlord/Warlord/Interfaces/Subsystems/GameApplication.cs
	Warlord/Warlord/View/FlyInput.cs
	Warlord/Warlord/Warlord.csproj
	Warlord/Warlord/WarlordApplication.cs
	Warlord/Warlord/bin/x86/Debug/Content/Fxs/block_effects.xnb
	Warlord/Warlord/bin/x86/Release/Content/Fxs/block_effects.xnb
	Warlord/Warlord/bin/x86/Release/Content/Fxs/skybox.xnb
	Warlord/Warlord/bin/x86/Release/Content/Fxs/warped_block_effects.xnb
	Warlord/WarlordContent/obj/x86/Debug/ContentPipeline.xml
	Warlord/WarlordContent/obj/x86/Release/ContentPipeline.xml

+++ Warlord/Warlord/GameTools/Noise3D/PerlinNoiseSettings3D.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-using Microsoft.Xna.Framework;
 using GameTools.Graph;
 
 namespace GameTools.Noise3D
@@ -14,44 +10,44 @@ namespace GameTools.Noise3D
 
         public Vector3i size;
         public Vector3i startingPoint;
-        public double frequencyMulti;
-        public double persistence;
-        public float zoom;        
-        public int octaves;      
+        public float frequencyMulti;
+        public float persistence;
+        public float zoom;
+        public int octaves;
         public int seed;
 
-        public PerlinNoiseSettings3D( )
+        public PerlinNoiseSettings3D()
         {
-            rng = new Random( );
+            rng = new Random();
 
-            size = new Vector3i( 100, 100, 100 );
-            startingPoint = Vector3i.Zero;  
-          
+            size = new Vector3i(100, 100, 100);
+            startingPoint = Vector3i.Zero;
+
             frequencyMulti = 2;
-            persistence = 0.5;
+            persistence = 0.5f;
             zoom = 40;
-            octaves = 6;            
-            seed = 0;   
+            octaves = 6;
+            seed = 0;
 
-            GenerateNewSeed( );
+            GenerateNewSeed();
         }
 
         public PerlinNoiseSettings3D(PerlinNoiseSettings3D settings)
         {
             size = new Vector3i(settings.size);
             startingPoint = new Vector3i(settings.startingPoint);
-          
+
             frequencyMulti = settings.frequencyMulti;
             persistence = settings.persistence;
             zoom = settings.zoom;
-            octaves = settings.octaves;            
-            seed = settings.seed;   
+            octaves = settings.octaves;
+            seed = settings.seed;
         }
 
-        public void GenerateNewSeed( )
+        public void GenerateNewSeed()
         {
-            seed = rng.Next( );
+            seed = rng.Next();
         }
-        
+
     }
 }
